+++ assets/src/scripts/components/daily-value-hydrograph/selectors/labels.js
@@ -1,6 +1,9 @@
 import {createSelector} from 'reselect';
 
-import {getCurrentDVTimeSeries} from '../../../selectors/daily-value-time-series-selector';
+import {
+    getCurrentDVTimeSeries,
+    getCurrentDVTimeSeriesUnitOfMeasure
+} from '../../../selectors/daily-value-time-series-selector';
 
 /*
  * Returns a selector function that returns a string that can be used as the title of the graph
@@ -8,8 +11,18 @@ import {getCurrentDVTimeSeries} from '../../../selectors/daily-value-time-series
 export const getCurrentTimeSeriesTitle = createSelector(
     getCurrentDVTimeSeries,
     (timeSeries) => {
-        return timeSeries && timeSeries.properties ?
-            timeSeries.properties.observedPropertyName : '';
+        // We will take the first available time series observedPropertyName
+        let result = '';
+        if (timeSeries) {
+            const propertyNames =
+                Object.values(timeSeries)
+                    .filter(ts => ts ? true : false)
+                    .map(ts => ts.properties.observedPropertyName);
+            if (propertyNames.length) {
+                result = propertyNames[0];
+            }
+        }
+        return result;
     }
 );
 
@@ -19,8 +32,23 @@ export const getCurrentTimeSeriesTitle = createSelector(
 export const getCurrentTimeSeriesDescription = createSelector(
     getCurrentDVTimeSeries,
     (timeSeries) => {
-        return timeSeries && timeSeries.properties ?
-            `${timeSeries.properties.observedPropertyName} for ${timeSeries.properties.samplingFeatureName}` : '';
+        // We will take the first available time series
+        let result = '';
+        if (timeSeries) {
+            const descriptions =
+                Object.values(timeSeries)
+                    .filter(ts => ts ? true : false)
+                    .map((ts) => {
+                        return {
+                            propertyName: ts.properties.observedPropertyName,
+                            featureName: ts.properties.samplingFeatureName
+                        };
+                    });
+            if (descriptions.length) {
+                result = `${descriptions[0].propertyName} for ${descriptions[0].featureName}`;
+            }
+        }
+        return result;
     }
 );
 
@@ -28,9 +56,10 @@ export const getCurrentTimeSeriesDescription = createSelector(
  * Returns a selector function that returns a string that can be used as the label of the y axis graph
  */
 export const getCurrentTimeSeriesYTitle = createSelector(
-    getCurrentDVTimeSeries,
-    (timeSeries) => {
-        return timeSeries && timeSeries.properties ?
-            `${timeSeries.properties.observedPropertyName}, ${timeSeries.properties.unitOfMeasureName}` : '';
+    getCurrentTimeSeriesTitle,
+    getCurrentDVTimeSeriesUnitOfMeasure,
+    (title, units) => {
+        return title && units ? `${title}, ${units}` : '';
+
     }
 );